// src/api/models/mint-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Immutable X API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.0
 * Contact: support@immutable.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { MintFee } from './mint-fee';
import { MintUser } from './mint-user';

/**
 * 
 * @export
 * @interface MintRequest
 */
export interface MintRequest {
    /**
     * Signature from authorised minter
     * @type {string}
     * @memberof MintRequest
     */
    'auth_signature': string;
    /**
     * minting contract
     * @type {string}
     * @memberof MintRequest
     */
    'contract_address': string;
    /**
     * Global contract-level royalty fees
     * @type {Array<MintFee>}
     * @memberof MintRequest
     */
    'royalties'?: Array<MintFee>;
    /**
     * Users to mint to
     * @type {Array<MintUser>}
     * @memberof MintRequest
     */
    'users': Array<MintUser>;
}